chore(sdk-coin-avaxc): move account-lib code
- moves code from account-lib to sdk-coin-avaxc

Ticket: BG-48144
