+++ modules/sdk-coin-avaxc/src/lib/index.ts
