+++ modules/sdk-coin-avaxc/src/lib/keyPair.ts
