+++ modules/sdk-coin-avaxc/src/lib/transactionBuilder.ts
@@ -1,13 +1,12 @@
 import { BaseCoin as CoinConfig } from '@bitgo/statics';
 import EthereumAbi from 'ethereumjs-abi';
-import { Eth } from '../../index';
 import { BuildTransactionError, TransactionType } from '@bitgo/sdk-core';
-import { Transaction } from '@bitgo/sdk-coin-eth';
+import { Transaction, TransactionBuilder as EthTransactionBuilder } from '@bitgo/sdk-coin-eth';
 import { walletSimpleByteCode, walletSimpleConstructor } from './walletUtil';
 import { getCommon } from './utils';
 import { TransferBuilder } from './transferBuilder';
 
-export class TransactionBuilder extends Eth.TransactionBuilder {
+export class TransactionBuilder extends EthTransactionBuilder {
   protected _transfer: TransferBuilder;
 
   constructor(_coinConfig: Readonly<CoinConfig>) {
