+++ modules/sdk-coin-avaxc/src/lib/utils.ts
