+++ modules/sdk-coin-avaxc/src/lib/walletUtil.ts
